+++ src/commands.rs
@@ -57,6 +57,12 @@ fn find_device(model: Option<args::DeviceModel>) -> anyhow::Result<nitrokey::Dev
   let device = iter
     .next()
     .with_context(|| format!("No Nitrokey device found{}", get_filter()))?;
+  anyhow::ensure!(
+    iter.next().is_none(),
+    "Multiple Nitrokey devices found{}.  Use the --model and --serial-number options to \
+    select one",
+    get_filter()
+  );
   Ok(device)
 }